requirements.txt
Flask==2.2.5
flask-restful==0.3.10
flask-mongoengine==1.0.0
flask-jwt-extended==4.5.3
marshmallow==3.20.1
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.1.0.post1